requirements.txt
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.1.2
